feat：http modify hertz

+++ internal/mi/builder/builder.go
@@ -12,8 +12,8 @@ type Builder struct {
 	RestrictedPackageName string            `json:"restricted_package_name,omitempty"`
 	Title                 string            `json:"title"`
 	Description           string            `json:"description"`
-	PassThrough           int32             `json:"pass_through"`          // 0 通知栏消息, 1 透传消息
-	NotifyType            int32             `json:"notify_type,omitempty"` // -1: DEFAULT_ALL 1: 使用默认提示音提示, 2: 使用默认震动提示, 4: 使用默认led灯光提示
+	PassThrough           int64             `json:"pass_through"`          // 0 通知栏消息, 1 透传消息
+	NotifyType            int64             `json:"notify_type,omitempty"` // -1: DEFAULT_ALL 1: 使用默认提示音提示, 2: 使用默认震动提示, 4: 使用默认led灯光提示
 	TimeToLive            int64             `json:"time_to_live,omitempty"`
 	TimeToSend            int64             `json:"time_to_send,omitempty"`
 	NotifyID              int64             `json:"notify_id,omitempty"`
